Now get_shot_file_path first searches clock in the DB.
I found the clock formatting to get the shot filename may differ between
systems. Which may impact code that manages historic shots. So if
get_shot_file_path is given a clock value, it now first searches it in
the database for the matching filename. If not found, then it just
formats the date.